Renamed a minor thing + Did func. tap the today panel to get a dialog that will set today to taken and record deviation from target med. taking time. Untested.

+++ app/src/main/java/com/example/andres/thirdypsinthrome/DayHolder.java
@@ -2,12 +2,12 @@ package com.example.andres.thirdypsinthrome;
 
 
 public class DayHolder{
-    int id;
+    long id;
     long date;
     float mg;
     boolean taken;
 
-    public DayHolder(int id, long date, float mg, int taken) {
+    public DayHolder(long id, long date, float mg, int taken) {
         this.id = id;
         this.date = date;
         this.mg = mg;
